Enable kernel64 and CI matrix

+++ console.c
@@ -55,7 +55,7 @@ void
 cprintf(char *fmt, ...)
 {
   int i, c, locking;
-  uint *argp;
+  unsigned long *argp;
   char *s;
 
   locking = cons.locking;
@@ -65,7 +65,7 @@ cprintf(char *fmt, ...)
   if (fmt == 0)
     panic("null fmt");
 
-  argp = (uint*)(void*)(&fmt + 1);
+  argp = (unsigned long*)(void*)(&fmt + 1);
   for(i = 0; (c = fmt[i] & 0xff) != 0; i++){
     if(c != '%'){
       consputc(c);
